Removed colors from Doctor pages, and added drag and drop feature
Introduced index to the array created from Rows.map. The index is used to keep track of positions for the elements in the row(table)
New functions introduced are: handleDragover, handleDrop, handleDragRow. These are used to manage positioning of the rows by acessing,
them through the index, and 'moving' them around in the row array